// src/api/models/TokenInfoDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TokenInfoDto = {
    accessToken?: string;
    id: number;
    login: string;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    permissions?: Array<string>;
};